refactor(api): move option code to own file
