+++ src/nvim/api/private/dispatch.c
@@ -24,6 +24,7 @@
 #include "nvim/api/autocmd.h"
 #include "nvim/api/buffer.h"
 #include "nvim/api/extmark.h"
+#include "nvim/api/options.h"
 #include "nvim/api/tabpage.h"
 #include "nvim/api/ui.h"
 #include "nvim/api/vim.h"
